# tests/contract/api-gateway-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent PPM Platform",
    "description": "AI-native Project Portfolio Management platform with 25 specialized agents",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/connectors": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "List Connectors",
        "description": "List all available connectors with their configuration status.\n\nOptionally filter by category.",
        "operationId": "list_connectors_v1_connectors_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectorListItemResponse"
                  },
                  "title": "Response List Connectors V1 Connectors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/connectors/{connector_id}": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "Get Connector",
        "description": "Get a specific connector with its configuration.",
        "operationId": "get_connector_v1_connectors__connector_id__get",
        "parameters": [
          {
            "name": "connector_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorListItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectorListItemResponse": {
        "properties": {
          "connector_id": {
            "type": "string",
            "title": "Connector Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "system": {
            "type": "string",
            "title": "System"
          },
          "mcp_server_id": {
            "type": "string",
            "title": "Mcp Server Id",
            "default": ""
          },
          "supported_operations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Operations"
          },
          "mcp_preferred": {
            "type": "boolean",
            "title": "Mcp Preferred"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "supported_sync_directions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Sync Directions"
          },
          "auth_type": {
            "type": "string",
            "title": "Auth Type"
          },
          "config_fields": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Config Fields"
          },
          "config_schema": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Config Schema"
          },
          "env_vars": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Env Vars"
          },
          "supported_objects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Objects"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "auth_requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Auth Requirements"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": false
          },
          "instance_url": {
            "type": "string",
            "title": "Instance Url",
            "default": ""
          },
          "project_key": {
            "type": "string",
            "title": "Project Key",
            "default": ""
          },
          "sync_direction": {
            "type": "string",
            "title": "Sync Direction",
            "default": "inbound"
          },
          "sync_frequency": {
            "type": "string",
            "title": "Sync Frequency",
            "default": "daily"
          },
          "health_status": {
            "type": "string",
            "title": "Health Status",
            "default": "unknown"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "custom_fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Custom Fields"
          },
          "mcp_server_url": {
            "type": "string",
            "title": "Mcp Server Url",
            "default": ""
          },
          "mcp_client_id": {
            "type": "string",
            "title": "Mcp Client Id",
            "default": ""
          },
          "mcp_client_secret": {
            "type": "string",
            "title": "Mcp Client Secret",
            "default": ""
          },
          "mcp_scope": {
            "type": "string",
            "title": "Mcp Scope",
            "default": ""
          },
          "mcp_scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mcp Scopes"
          },
          "mcp_api_key": {
            "type": "string",
            "title": "Mcp Api Key",
            "default": ""
          },
          "mcp_api_key_header": {
            "type": "string",
            "title": "Mcp Api Key Header",
            "default": ""
          },
          "mcp_oauth_token": {
            "type": "string",
            "title": "Mcp Oauth Token",
            "default": ""
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "default": ""
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret",
            "default": ""
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "mcp_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mcp Tools"
          },
          "mcp_tool_map": {
            "additionalProperties": true,
            "type": "object",
            "title": "Mcp Tool Map"
          },
          "prefer_mcp": {
            "type": "boolean",
            "title": "Prefer Mcp",
            "default": false
          },
          "mcp_enabled": {
            "type": "boolean",
            "title": "Mcp Enabled",
            "default": true
          },
          "mcp_feature_enabled": {
            "type": "boolean",
            "title": "Mcp Feature Enabled",
            "default": true
          },
          "mcp_enabled_operations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mcp Enabled Operations"
          },
          "mcp_disabled_operations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mcp Disabled Operations"
          }
        },
        "type": "object",
        "required": [
          "connector_id",
          "name",
          "description",
          "category",
          "system",
          "supported_operations",
          "mcp_preferred",
          "status",
          "icon",
          "supported_sync_directions",
          "auth_type",
          "config_fields",
          "config_schema",
          "env_vars"
        ],
        "title": "ConnectorListItemResponse",
        "description": "Response model for connector list items (combines definition and config)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}